+++ eggs/Products.NaayaGlossary/Products/NaayaGlossary/NyGlossary.py
@@ -739,6 +739,17 @@ class NyGlossary(Folder, utils, catalog_utils, glossary_export, file_utils):
     security.declareProtected(PERMISSION_MANAGE_NAAYAGLOSSARY,
                               'updateGlossaryProperties')
 
+    security.declareProtected(PERMISSION_MANAGE_NAAYAGLOSSARY, 'del_element')
+
+    def del_element(self, REQUEST):
+        """ delete a glossary element """
+        folder_id, element_id = REQUEST.get('item').split('/')
+        folder = getattr(self, folder_id)
+        if element_id:
+            folder.manage_delObjects(element_id)
+            self.setSessionInfoTrans('Element successfully deleted')
+        return REQUEST.RESPONSE.redirect(self.absolute_url())
+
     def updateGlossaryProperties(self, title='', approved='',
                                  parent_anchors='', REQUEST=None):
         """ """
@@ -794,7 +805,10 @@ class NyGlossary(Folder, utils, catalog_utils, glossary_export, file_utils):
             parent.manage_addGlossaryElement(id, entry_title, entry_subjects,
                                              entry_source, entry_contributor)
         if REQUEST:
-            return REQUEST.RESPONSE.redirect('index_html')
+            if REQUEST.get('redirect_to_folder'):
+                return REQUEST.RESPONSE.redirect(parent.absolute_url())
+            else:
+                return REQUEST.RESPONSE.redirect('index_html')
 
     def checkPermissionManageGlossary(self):
         return getSecurityManager().checkPermission(